\name{is.polymars}
\alias{is.polymars}
\title{
test for polymars objects
}
\usage{
is.polymars(object)
}
\arguments{
\item{object}{
the object under investigation.
}}
\value{
returns TRUE if the object is a polymars model, FALSE otherwise.
}
\seealso{
plot.polymars, predict.polymars, summary.polymars, print.polymars


}
% Converted by Sd2Rd version 0.2-a4.


