\name{is.polymars}
\alias{is.polymars}
\title{
test for polymars objects
}
\description{test for objects of class \code{"polymars"}.}
\usage{
is.polymars(object)
}
\arguments{
\item{object}{
the object under investigation.
}}
\value{
returns \code{TRUE} if the object is a polymars model, \code{FALSE} otherwise.
}
\seealso{
  \code{\link{plot.polymars}}, \code{\link{predict.polymars}},
  \code{\link{summary.polymars}}, \code{\link{print.polymars}}
}
% Converted by Sd2Rd version 0.2-a4.


